Move ServerOptions to IceRpc.Configure
