+++ examples/GenericHost/Server/appsettings.json
@@ -9,7 +9,7 @@
   },
   "Server": {
     "Endpoint": "icerpc://127.0.0.1:10000",
-    "CloseTimeout": "00:00:05",
+    "CloseTimeout": "00:00:05"
   },
   "Transport": {
     "CertificateFile": "server.p12",
